remove hardcoded file size limit by getting the file size dynamically

+++ charkernel.cl
@@ -1,6 +1,5 @@
-__kernel 
-void myKernel(__global char *myChar, __global int *B) {
-	int i = get_global_id(0);
+__kernel void myKernel(__global char *myChar, __global int *B) {
+	size_t i = get_global_id(0);
 	if (myChar[i] == 'e') {
 		atomic_add(B, 1);
 	}
